--- PORTFOLIO/Get_data.py
@@ -1,6 +1,0 @@
-import requests
-import csv
-import MPT_Module as mptm
-
-mptm.get_price_data()
-# 시간이 오래걸려서 따로 빼둠.
