+++ SpringStudy/PortfolioMPT/optimization.py
@@ -1,0 +1,9 @@
+import pandas as pd
+import csv
+import numpy as np
+import MPT_Module as mptm
+import matplotlib.pyplot as plt
+import scipy.optimize as sco
+
+print(1)
+mptm.stat_Port_Opi_plot()
